pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name        = "cerberus-go2"
version     = "3.1.0"
description = "Canine-Emulative Responsive Behavioral Engine & Reactive Utility System for Unitree Go2"
readme      = "Readme.md"
license     = { text = "MIT" }
requires-python = ">=3.11"
authors     = [{ name = "CERBERUS Contributors" }]
keywords    = ["robotics", "unitree", "go2", "quadruped", "ai", "behavior"]

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dds     = ["unitree_sdk2py>=1.0.0"]
webrtc  = ["go2_webrtc_connect>=1.0.0"]
vision  = ["ultralytics>=8.2.0", "mediapipe>=0.10.0", "opencv-python>=4.9.0"]
audio   = ["PyAudio>=0.2.14", "SpeechRecognition>=3.10.0"]
dev     = ["pytest>=8.1.0", "pytest-asyncio>=0.23.0", "httpx>=0.27.0", "ruff>=0.4.0"]
all     = ["cerberus-go2[dds,webrtc,vision,audio,dev]"]

[project.scripts]
cerberus        = "cerberus.cli:main"
cerberus-server = "backend.api.server:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length    = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501", "B008", "SIM118"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]
